Add ArtifactFromDir to the test server
