+++ http/fetch/archive_fetcher_test.go
@@ -17,14 +17,9 @@ limitations under the License.
 package fetch
 
 import (
-	"archive/tar"
-	"compress/gzip"
-	"crypto/sha256"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	. "github.com/onsi/gomega"
@@ -46,7 +41,7 @@ func TestArchiveFetcher_Fetch(t *testing.T) {
 	fileName := "testdata/manifests.yaml"
 	artifactName := "manifests.tgz"
 	artifactURL := fmt.Sprintf("%s/%s", testServer.URL(), artifactName)
-	artifactChecksum, err := createArtifact(testServer, "testdata", artifactName)
+	artifactChecksum, err := testServer.ArtifactFromDir("testdata", artifactName)
 	g.Expect(err).ToNot(HaveOccurred())
 
 	tests := []struct {
@@ -103,93 +98,4 @@ func TestArchiveFetcher_Fetch(t *testing.T) {
 		})
 	}
 
-}
-
-func createArtifact(artifactServer *testserver.ArtifactServer, source, destination string) (string, error) {
-	if f, err := os.Stat(source); os.IsNotExist(err) || !f.IsDir() {
-		return "", fmt.Errorf("invalid source path: %s", source)
-	}
-	f, err := os.Create(filepath.Join(artifactServer.Root(), destination))
-	if err != nil {
-		return "", err
-	}
-	defer func() {
-		if err != nil {
-			os.Remove(f.Name())
-		}
-	}()
-
-	h := sha256.New()
-
-	mw := io.MultiWriter(h, f)
-	gw := gzip.NewWriter(mw)
-	tw := tar.NewWriter(gw)
-
-	if err = filepath.Walk(source, func(p string, fi os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Ignore anything that is not a file (directories, symlinks)
-		if !fi.Mode().IsRegular() {
-			return nil
-		}
-
-		// Ignore dotfiles
-		if strings.HasPrefix(fi.Name(), ".") {
-			return nil
-		}
-
-		header, err := tar.FileInfoHeader(fi, p)
-		if err != nil {
-			return err
-		}
-		// The name needs to be modified to maintain directory structure
-		// as tar.FileInfoHeader only has access to the base name of the file.
-		// Ref: https://golang.org/src/archive/tar/common.go?#L626
-		relFilePath := p
-		if filepath.IsAbs(source) {
-			relFilePath, err = filepath.Rel(source, p)
-			if err != nil {
-				return err
-			}
-		}
-		header.Name = relFilePath
-
-		if err := tw.WriteHeader(header); err != nil {
-			return err
-		}
-
-		f, err := os.Open(p)
-		if err != nil {
-			f.Close()
-			return err
-		}
-		if _, err := io.Copy(tw, f); err != nil {
-			f.Close()
-			return err
-		}
-		return f.Close()
-	}); err != nil {
-		return "", err
-	}
-
-	if err := tw.Close(); err != nil {
-		gw.Close()
-		f.Close()
-		return "", err
-	}
-	if err := gw.Close(); err != nil {
-		f.Close()
-		return "", err
-	}
-	if err := f.Close(); err != nil {
-		return "", err
-	}
-
-	if err := os.Chmod(f.Name(), 0644); err != nil {
-		return "", err
-	}
-
-	return fmt.Sprintf("%x", h.Sum(nil)), nil
 }
